Made MAJOR changes to how Feature objects har made
Previously we only called a static method in Feature. Now we must create
a Feature object instance and specify which single features should be
used in that specific instance. When all are chosen, the Feature object is
set as finished, and only then can a featureVector be generated and
returned.